refactor(muntarfs): move archive implementation into library
