+++ AGENTS.md
@@ -24,6 +24,9 @@ Main files:
 - [src/cjit.h](/home/jrml/devel/cjit/src/cjit.h): `CJITState` and public runtime functions
 - [src/file.c](/home/jrml/devel/cjit/src/file.c): file/stdin/path helpers
 - [lib/muntarfs/muntarfs.h](/home/jrml/devel/cjit/lib/muntarfs/muntarfs.h): bundle extraction surface used by CJIT
+- [lib/muntarfs/muntar.c](/home/jrml/devel/cjit/lib/muntarfs/muntar.c): tar extraction and archive reader
+- [lib/muntarfs/tinflate.c](/home/jrml/devel/cjit/lib/muntarfs/tinflate.c): low-level deflate implementation
+- [lib/muntarfs/tinfgzip.c](/home/jrml/devel/cjit/lib/muntarfs/tinfgzip.c): gzip wrapper over the inflater
 - [src/elflinker.c](/home/jrml/devel/cjit/src/elflinker.c): POSIX library resolution
 - [src/winlinker.c](/home/jrml/devel/cjit/src/winlinker.c): Windows library resolution
 - [src/win-compat.c](/home/jrml/devel/cjit/src/win-compat.c): Windows compatibility helpers
@@ -82,7 +85,7 @@ Use this direction for refactors:
 - `src/adapters/fs/`: file/path/tempdir/asset/archive IO
 - `src/adapters/platform/`: process and platform-specific library resolution
 - `src/support/`: small low-level reusable support only
-- `lib/muntarfs/`: standalone bundle pack/extract component for embedded tar/tar.gz assets
+- `lib/muntarfs/`: standalone bundle pack/extract component and archive implementation for embedded tar/tar.gz assets
 
 Current code is not fully there yet. New changes should avoid making `src/main.c` and `src/cjit.c` even broader.
 
@@ -150,8 +153,9 @@ Current start point:
 
 - [src/file.c](/home/jrml/devel/cjit/src/file.c)
 - [src/cjit.c](/home/jrml/devel/cjit/src/cjit.c)
-- [src/tinflate.c](/home/jrml/devel/cjit/src/tinflate.c)
-- [src/tinfgzip.c](/home/jrml/devel/cjit/src/tinfgzip.c)
+- [lib/muntarfs/muntar.c](/home/jrml/devel/cjit/lib/muntarfs/muntar.c)
+- [lib/muntarfs/tinflate.c](/home/jrml/devel/cjit/lib/muntarfs/tinflate.c)
+- [lib/muntarfs/tinfgzip.c](/home/jrml/devel/cjit/lib/muntarfs/tinfgzip.c)
 - [lib/muntarfs/muntarfs_runtime.c](/home/jrml/devel/cjit/lib/muntarfs/muntarfs_runtime.c)
 
 Target destination:
